move prayer times providers to shared providers
