+++ lib/pages/prayer_time_page/prayer_time.dart
@@ -4,12 +4,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import 'package:qurantafsir_flutter/pages/prayer_time_page/prayer_times_notifier.dart';
-
 import 'package:qurantafsir_flutter/shared/constants/image.dart';
 import 'package:qurantafsir_flutter/shared/constants/prayer_times_list.dart';
 import 'package:qurantafsir_flutter/shared/constants/qp_colors.dart';
 import 'package:qurantafsir_flutter/shared/constants/qp_text_style.dart';
+import 'package:qurantafsir_flutter/shared/core/providers/prayer_times_notifier.dart';
 import 'package:qurantafsir_flutter/shared/ui/state_notifier_connector.dart';
 
 class PrayerTimePage extends ConsumerStatefulWidget {
